webapp/requirements.txt
# --- Estos son comentarios---
# Para instalar una versión específica
# package==version

# Para instalar una versión igual o superior
# package>=version

# Para instalar la versión más reciente
# package

# Para usar el archivo
# pip install -r requirements.txt
# pip freeze lista todos los paquetes instalados y su version
# -----------------------

flask
flask_restful
requests
mysql-connector-python
